Port OrdRec to new API
